+++ recommendation/service/recommendation_repository.py
@@ -1,7 +1,7 @@
 from utils.storage.s3_manager import S3Manager
 from utils.storage.opensearch_manager import OpenSearchManager
 from recommendation.transformer.recommendation_transformer import RecommendationTransformer
-from config.recommendation_config import OPENSEARCH_CONFIG
+from config.opensearch_config import OPENSEARCH_CONFIG
 from typing import Dict, Any, Optional
 import logging
 logger = logging.getLogger(__name__)
